auth: extract popup/redirect/credential flow-staging into AuthFlowRegistry
Move the resolver/mock/redirect-result island out of the God-file into
sandbox-auth-flow.ts as a self-contained AuthFlowRegistry. This state
(one-shot mockResults, the injected resolver, the pending redirectResult
slot) is provably independent of the coupled auth state machine — no
other backend method reads it — so it composes cleanly: the backend holds
one 'flow' field and the six accessor methods delegate verbatim. The
public surface is unchanged (the six methods stay on SandboxBackend).

Deliberately sized as the seam for the imminent email-link / linking /
reauth climb: the 'stage a result, resolve a flow, hand back a
credential' pattern those flows extend lives here, with the interface
documented in the file header. Identity mutation stays on the core.
